major refactor splitting ingest for arrays from exomes/genomes
