+++ src/main/java/org/broadinstitute/hellbender/tools/variantdb/ingest/IngestConstants.java
@@ -1,0 +1,10 @@
+package org.broadinstitute.hellbender.tools.variantdb.ingest;
+
+public class IngestConstants {
+
+    public static final char SEPARATOR = '\t';
+    public static final String FILETYPE = ".tsv";
+    public static final String metadataDirectoryName = "metadata";
+    public static final int partitionPerTable = 4000;
+
+}
